KITE-1023: Fix PartitionView URI escaping.
This adds test for URIs with escaped characters and fixes handling in
the FS PartitionView implementation. The implementation was accidentally
unescaping URIs by using URI.create, which expects escaped characters,
along with URI#getPath, which will remove escapes.

This also fixes the URI returned by getLocation. This previously used
Path#getUri, but that method returns the Path's internal URI that
is double-escaped. Returning a URI created from Path#toString is the
correct behavior because it matches the Path passed in. Similarly,
the merge and replace methods in the FS dataset implementation that rely
on Paths created from those locations have been updated to correctly
construct a Path from a URI by converting the URI to a String rather
than setting the internal URI directly.